+++ packages/xrpl/src/models/ledger/NFTokenPage.ts
@@ -1,5 +1,3 @@
-import { ModifiedNode, Node } from '../transactions/metadata'
-
 import BaseLedgerEntry from './BaseLedgerEntry'
 
 export interface NFToken {
@@ -21,10 +19,4 @@ export interface NFTokenPage extends BaseLedgerEntry {
   PreviousPageMin?: string
   PreviousTxnID?: string
   PreviousTxnLgrSeq?: number
-}
-
-export function isNFTokenPage(
-  ledgerEntry: BaseLedgerEntry,
-): ledgerEntry is NFTokenPage {
-  return Object.prototype.hasOwnProperty.call(ledgerEntry, `NFTokens`)
 }